+++ test/queue-message.unit.ts
@@ -23,14 +23,12 @@ describe('QueueMessage', () => {
             message: 'Sample message',
             messageId: '123',
             messageType: 'sample-event',
-            publishedDate: ('{publishedDate}'),
-            file_size_mb: 1.5
+            publishedDate: ('{publishedDate}')
         });
         // Assert
         expect(queuemessage.message).toBe("Sample message");
         expect(queuemessage.messageId).toBe("123");
         expect(queuemessage.messageType).toBe("sample-event");
-        expect(queuemessage.fileSizeMb).toBe(1.5);
         // expect(queuemessage.publishedDate).toBe(publishedDate);
     });
     it('Should Instantiate', () => {
@@ -38,16 +36,6 @@ describe('QueueMessage', () => {
         const queueMessage = QueueMessage.from();
         // Assert
         expect(queueMessage).toBeInstanceOf(QueueMessage);
-    });
-
-    it('Should convert file_size_mb string to number', () => {
-        // Arrange
-        const queueMessage = QueueMessage.from({
-            file_size_mb: '20'
-        });
-        // Assert
-        expect(queueMessage.fileSizeMb).toBe(20);
-        expect(typeof queueMessage.fileSizeMb).toBe("number");
     });
 
 })